core/directives: split up directives into separate files
