+++ scripts/core/directives/SelectDirective.js
@@ -240,29 +240,4 @@ function ($parse, $document, $compile, optionParser) {
             };
         }
     };
-}])
-
-.directive('sdSelectPopup', ['$document', 'asset', function ($document, asset) {
-    return {
-        restrict: 'A',
-        scope: false,
-        replace: true,
-        templateUrl: asset.templateUrl('core/views/sdselect.html'),
-        link: function (scope, element, attrs) {
-
-            scope.open = false;
-
-            scope.toggleSelect = function () {
-                scope.open = !scope.open;
-                if (scope.open) {
-                    scope.focus();
-                }
-            };
-
-            scope.focus = function focus() {
-                var searchBox = element.find('input')[0];
-                searchBox.focus();
-            };
-        }
-    };
 }]);
