+++ scripts/core/directives/index.js
@@ -1,14 +1,22 @@
 import './AutofocusDirective';
-import './DebounceThrottleDirective';
+import './ThrottleDirective';
+import './DebounceDirective';
 import './SortDirective';
 import './PasswordStrengthDirective';
 import './SearchListDirective';
+import './SearchListSingleDirective';
 import './FiletypeIconDirective';
 import './CheckDirective';
+import './CheckAllDirective';
+import './SwitchDirective';
+import './SwitchInvertedDirective';
 import './ConfirmDirective';
 import './SelectDirective';
+import './SelectPopupDirective';
 import './PermissionsDirective';
-import './DragDropDirective';
+import './SortableDirective';
+import './DroppableDirective';
+import './DraggableDirective';
 import './TypeaheadDirective';
 import './SliderDirective';
 import './UserAvatarDirective';
@@ -19,21 +27,29 @@ import './WithParamsDirective';
  * @module superdesk.core.directives
  * @name superdesk.core.directives
  * @packageName superdesk.core
- * @description Superdesk core directives collection. Contains a set of modules 
- * that implement various UI components and helpers.
+ * @description Superdesk core directives collection. Contains a set of modules
+ *  that implement various UI components and helpers.
  */
 export default angular.module('superdesk.core.directives', [
     'superdesk.core.directives.autofocus',
+    'superdesk.core.directives.debounce',
     'superdesk.core.directives.throttle',
     'superdesk.core.directives.sort',
     'superdesk.core.directives.passwordStrength',
     'superdesk.core.directives.searchList',
+    'superdesk.core.directives.searchListSingle',
     'superdesk.core.directives.filetypeIcon',
     'superdesk.core.directives.check',
+    'superdesk.core.directives.checkAll',
+    'superdesk.core.directives.switch',
+    'superdesk.core.directives.switchInverted',
     'superdesk.core.directives.confirm',
     'superdesk.core.directives.select',
+    'superdesk.core.directives.selectPopup',
     'superdesk.core.directives.permissions',
-    'superdesk.core.directives.dragdrop',
+    'superdesk.core.directives.sortable',
+    'superdesk.core.directives.draggable',
+    'superdesk.core.directives.droppable',
     'superdesk.core.directives.typeahead',
     'superdesk.core.directives.slider',
     'superdesk.core.directives.avatar',
